refactor(dashboard): organize metric files & address PR comments
