+++ dashboard/src/app/(authenticated)/[tenantId]/(diagram)/wfSpec/[...props]/components/metrics/index.ts
@@ -1,0 +1,1 @@
+export { WfSpecMetrics } from './WfSpecMetrics'
