+++ dashboard/src/app/(authenticated)/[tenantId]/(diagram)/wfSpec/[...props]/components/metrics/wfSpecMetricsTypes.ts
@@ -1,0 +1,1 @@
+export type ViewMode = 'count' | 'latency'
